requirements.txt
streamlit
groq
pdfplumber
playwright
beautifulsoup4
lxml
youtube-transcript-api
pytesseract
pillow
